refactor: 拆分条目解析模块
